deduplicates cart resolve logic
