+++ packages/api/resolvers/mutations/addCartProduct.js
@@ -1,13 +1,7 @@
 import { log } from 'meteor/unchained:core-logger';
 import { Products } from 'meteor/unchained:core-products';
-import { Users } from 'meteor/unchained:core-users';
-import { Orders } from 'meteor/unchained:core-orders';
-import {
-  ProductNotFoundError,
-  UserNotFoundError,
-  OrderNotFoundError,
-  OrderWrongStatusError
-} from '../../errors';
+import { ProductNotFoundError } from '../../errors';
+import getCart from '../../getCart';
 
 export default function(
   root,
@@ -22,21 +16,7 @@ export default function(
   );
   const product = Products.findOne({ _id: productId });
   if (!product) throw new ProductNotFoundError({ data: { productId } });
-  if (orderId) {
-    const order = Orders.findOne({ _id: orderId });
-    if (!order) throw new OrderNotFoundError({ orderId });
-    if (!order.isCart()) {
-      throw new OrderWrongStatusError({ data: { status: order.status } });
-    }
-    return order.addProductItem({
-      product,
-      quantity,
-      configuration
-    });
-  }
-  const user = Users.findOne({ _id: userId });
-  if (!user) throw new UserNotFoundError({ userId });
-  const cart = user.initCart({ countryContext });
+  const cart = getCart({ orderId, userId, countryContext });
   return cart.addProductItem({
     product,
     quantity,
