+++ packages/api/resolvers/mutations/checkoutCart.js
@@ -1,28 +1,6 @@
 import { log } from 'meteor/unchained:core-logger';
-import { Orders } from 'meteor/unchained:core-orders';
-import { Users } from 'meteor/unchained:core-users';
-import {
-  UserNotFoundError,
-  OrderCheckoutError,
-  UserNoCartError,
-  OrderWrongStatusError,
-  OrderNotFoundError
-} from '../../errors';
-
-const checkoutWithPotentialErrors = (cart, context, options, userId) => {
-  try {
-    return cart.checkout(context, options);
-  } catch (error) {
-    const data = {
-      userId,
-      orderId: cart._id,
-      ...context,
-      detailMessage: error.message
-    };
-    log(data.detailMessage, { userId, orderId: cart._id, level: 'error' });
-    throw new OrderCheckoutError({ data });
-  }
-};
+import { OrderCheckoutError } from '../../errors';
+import getCart from '../../getCart';
 
 export default function(
   root,
@@ -30,31 +8,19 @@ export default function(
   { userId, countryContext, localeContext }
 ) {
   log('mutation checkoutCart', { orderId, userId });
-  if (orderId) {
-    const order = Orders.findOne({ _id: orderId });
-    if (!order) throw new OrderNotFoundError({ data: { orderId } });
-    if (!order.isCart()) {
-      throw new OrderWrongStatusError({ data: { status: order.status } });
-    }
-    return checkoutWithPotentialErrors(
-      order,
-      transactionContext,
-      {
-        localeContext
-      },
-      userId
-    );
-  }
-  const user = Users.findOne({ _id: userId });
-  if (!user) throw new UserNotFoundError({ data: { userId } });
-  const cart = user.cart({ countryContext });
-  if (!cart) throw new UserNoCartError({ data: { userId } });
-  return checkoutWithPotentialErrors(
-    cart,
-    transactionContext,
-    {
+  const cart = getCart({ orderId, userId, countryContext });
+  try {
+    return cart.checkout(transactionContext, {
       localeContext
-    },
-    userId
-  );
+    });
+  } catch (error) {
+    const data = {
+      userId,
+      orderId: cart._id,
+      ...transactionContext,
+      detailMessage: error.message
+    };
+    log(data.detailMessage, { userId, orderId: cart._id, level: 'error' });
+    throw new OrderCheckoutError({ data });
+  }
 }
