+++ packages/api/resolvers/mutations/updateCart.js
@@ -1,11 +1,5 @@
 import { log } from 'meteor/unchained:core-logger';
-import { Users } from 'meteor/unchained:core-users';
-import { Orders } from 'meteor/unchained:core-orders';
-import {
-  UserNotFoundError,
-  OrderNotFoundError,
-  OrderWrongStatusError
-} from '../../errors';
+import getCart from '../../getCart';
 
 export default function(
   root,
@@ -13,19 +7,7 @@ export default function(
   { countryContext, userId }
 ) {
   log('mutation updateCart', { userId });
-  let order;
-  if (orderId) {
-    order = Orders.findOne({ _id: orderId });
-    if (!order) throw new OrderNotFoundError({ data: { orderId } });
-    if (!order.isCart()) {
-      throw new OrderWrongStatusError({ data: { status: order.status } });
-    }
-  } else {
-    const user = Users.findOne({ _id: userId });
-    if (!user) throw new UserNotFoundError({ userId });
-    order = user.initCart({ countryContext });
-  }
-
+  let order = getCart({ orderId, userId, countryContext });
   if (meta) {
     order = order.updateContext(meta);
   }
